pyproject.toml
[tool.poetry]
name = "azure-postgre-sqlalchemy"
version = "0.1.0"
description = ""
authors = ["bn-pam <bonneau.pam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.31"
sqlalchemy-orm = "^1.2.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
